Module storage_proofs_post::fallback

Structs

ChallengeRequirements

ComponentPrivateInputs

FallbackPoSt

FallbackPoStCircuit

This is the FallbackPoSt circuit.

FallbackPoStCompound

PrivateInputs

PrivateSector

Proof

PublicInputs

PublicParams

PublicSector

Sector

SectorProof

SetupParams

Functions

generate_leaf_challenge

Generates challenge, such that the range fits into the sector.

generate_leaf_challenge_inner

generate_leaf_challenges

Generate all challenged leaf ranges for a single sector, such that the range fits into the sector.

generate_sector_challenge

Generate a single sector challenge.

generate_sector_challenges

get_challenge_index

Selects the challenge index used to determine the leaf challenge for PoSt

vanilla_proof